feat: Add Item and Order models
This commit adds the Item and Order models to the application. The Item model represents an item in an order and has attributes such as quantity, price, order ID, and product ID. It also includes relationships with the Order and Product models. The Order model represents an order and has attributes such as total and user ID. It also includes a relationship with the User model. These models will be used to manage orders and items in the application.